prevent writing a file already written

+++ photo.py
@@ -42,8 +42,11 @@ def save(frame, visitation_id, detection_score, photo_type):
         if has_disk_space():
             directory = mkdirs(visitation_id)
             image_path = "{}/{}_{}_{}.png".format(directory, photo_type, time.strftime("%H-%M-%S"), detection_score)
-            logging.info("writing image {}".format(image_path))
-            cv2.imwrite( image_path, frame )
+            if not path.isfile(image_path):
+                logging.info("writing image {}".format(image_path))
+                cv2.imwrite( image_path, frame )
+            else:
+                logging.info("file already exists")
             # camera = PiCamera()
             # camera.resolution = (3264, 2448)
             # if bounding_box:
